+++ pytsite/google/res/assets/js/common.js
@@ -1,0 +1,1 @@
+pytsite.google = {};