+++ pytsite/google/res/assets/js/pytsite-google.js
@@ -1,8 +1,0 @@
-pytsite.google = {
-    ready: false,
-
-    initCallback: function() {
-        $(window).trigger('pytsite.google.ready');
-        pytsite.google.ready = true;
-    }
-};