+++ pytsite/reddit/res/assets/js/widget.js
@@ -1,4 +1,0 @@
-$(function () {
-    $('.widget-reddit-oauth').each(function () {
-    });
-});